requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Document Processing
docling==1.0.0
python-docx==1.1.0
PyPDF2==3.0.1
markdown==3.5.1

# RAG Framework
llama-index==0.9.20
llama-index-vector-stores-postgres==0.1.3
llama-index-embeddings-huggingface==0.1.3
llama-index-llms-ollama==0.1.2

# Vector Database
pgvector==0.2.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Embeddings and Re-ranking
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.0

# Agentic Framework
crewai==0.1.0
langchain==0.0.350

# LLM Hosting
ollama==0.1.6

# Observability
arize-phoenix==2.0.0
openinference-instrumentation-llama-index==0.1.0

# Evaluation
ragas==0.1.0
datasets==2.14.7

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiohttp==3.9.1
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

